+++ NEOFORGE/src/main/java/com/createcompletelycreate/CreateCompletelyCreate.java
@@ -15,9 +15,11 @@ import com.simibubi.create.foundation.item.TooltipModifier;
 import net.createmod.catnip.lang.FontHelper;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.ModList;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
-import net.neoforged.fml.ModList;
+import net.neoforged.neoforge.client.gui.ConfigurationScreen;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -48,6 +50,9 @@ public class CreateCompletelyCreate {
         ModCreativeTabs.register(modEventBus);
         ModConfigs.register(modContainer);
         ModConfigs.registerConditionCodecs(modEventBus);
+
+        modContainer.registerExtensionPoint(IConfigScreenFactory.class,
+                (minecraft, parent) -> new ConfigurationScreen(modContainer, parent));
 
         ModRecipes.register(modEventBus);
         ModRecipeRequirementTypes.init();