+++ NEOFORGE/src/main/java/com/createcompletelycreate/components/extruder/brass/BrassExtruderBlockEntity.java
@@ -3,7 +3,6 @@ package com.createcompletelycreate.components.extruder.brass;
 import com.createcompletelycreate.components.extruder.AbstractExtruderBlockEntity;
 import com.createcompletelycreate.register.ModBlockEntities;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.neoforged.neoforge.capabilities.Capabilities;
@@ -23,8 +22,7 @@ public class BrassExtruderBlockEntity extends AbstractExtruderBlockEntity {
         event.registerBlockEntity(
                 Capabilities.ItemHandler.BLOCK,
                 ModBlockEntities.BRASS_BLOCK_EXPELLER.get(),
-                (be, context) -> context == Direction.DOWN ? be.getItemHandler() : null
+                (be, context) -> be.getItemHandler()
         );
-
     }
 }